refactor: isolate auto post scheduler
